requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.0.8
django-allauth==0.42.0
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.11.1
drf-yasg==1.20.0
entrypoints==0.3
flake8==3.7.9
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
pyflakes==2.1.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2019.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
